pyproject.toml
[project]
name = "vcr"
version = "0.1.0"
description = "Vehicle Color Recognition - Multi-Scale CNN Classifier"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "torchinfo>=1.8.0",
    "mlflow>=2.10.0",
    "ultralytics>=8.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "cvat-sdk>=2.0.0",
    "optuna>=3.5.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
